+++ org.jenetics/src/test/java/org/jenetics/Integer64GeneTest.java
@@ -137,7 +137,7 @@ public class Integer64GeneTest extends NumberGeneTester<Integer64, Integer64Gene
 			assertEquals(b.getMax().longValue(), max);
 			assertEquals(c.getMin().longValue(), min);
 			assertEquals(c.getMax().longValue(), max);
-			assertEquals(c.getAllele().longValue(), (long)(((i - 50) + ((i - 100)*3))/2));
+			assertEquals(c.getAllele().longValue(), ((i - 50) + ((i - 100)*3))/2);
 		}
 	}
